chore (refs T22614): move DpBulkEditHeader to client/js.
