+++ client/js/components/core/DpBulkEditHeader.vue
