+++ demosplan/DemosPlanNewsBundle/Resources/client/js/components/DpNewsAdminList.vue
@@ -75,7 +75,7 @@
 
 <script>
 import { dpApi } from '@DemosPlanCoreBundle/plugins/DpApi'
-import DpBulkEditHeader from '@DemosPlanCoreBundle/components/DpBulkEditHeader'
+import DpBulkEditHeader from '@DpJs/components/core/DpBulkEditHeader'
 import DpDataTable from '@DpJs/components/core/DpDataTable/DpDataTable'
 import DpNewsItemStatus from './DpNewsItemStatus'
 import { makeFormPost } from '@DpJs/lib/utils/makeFormPost'
