+++ demosplan/DemosPlanProcedureBundle/Resources/client/js/components/SegmentsList/SegmentsList.vue
@@ -232,7 +232,7 @@ import { CleanHtml, VPopover } from 'demosplan-ui/directives'
 import { DpButton, DpLoading } from 'demosplan-ui/components'
 import { mapActions, mapGetters, mapMutations, mapState } from 'vuex'
 import CustomSearch from './CustomSearch'
-import DpBulkEditHeader from '@DemosPlanCoreBundle/components/DpBulkEditHeader'
+import DpBulkEditHeader from '@DpJs/components/core/DpBulkEditHeader'
 import DpColumnSelector from '@DpJs/components/core/DpDataTable/DpColumnSelector'
 import DpDataTable from '@DpJs/components/core/DpDataTable/DpDataTable'
 import DpFlyout from '@DpJs/components/core/DpFlyout'
