+++ demosplan/DemosPlanStatementBundle/Resources/client/js/components/listStatements/ListStatements.vue
@@ -275,7 +275,7 @@ import { checkResponse, dpApi, dpRpc } from '@DemosPlanCoreBundle/plugins/DpApi'
 import { DpButton, DpLoading } from 'demosplan-ui/components'
 import { mapActions, mapMutations, mapState } from 'vuex'
 import { CleanHtml } from 'demosplan-ui/directives'
-import DpBulkEditHeader from '@DemosPlanCoreBundle/components/DpBulkEditHeader'
+import DpBulkEditHeader from '@DpJs/components/core/DpBulkEditHeader'
 import DpClaim from '@DemosPlanStatementBundle/components/DpClaim'
 import DpDataTable from '@DpJs/components/core/DpDataTable/DpDataTable'
 import DpFlyout from '@DpJs/components/core/DpFlyout'
